requirements.txt
fastapi
uvicorn
python-dotenv
requests
openai
pydantic
watchdog
sncl
python-multipart
typing-extensions
agency-swarm
nest_asyncio
certifi
pandas
jupyter
redis
aiohttp
pytz
qrcode
opencv-python
numpy==1.26